requirements.txt
# Autonomous Content Machine Dependencies
openai>=1.0.0
anthropic>=0.18.0
pyyaml>=6.0
jinja2>=3.1.0
markdown>=3.5.0
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
schedule>=1.2.0
python-dotenv>=1.0.0
rich>=13.0.0
click>=8.1.0
aiohttp>=3.9.0
aiofiles>=23.0.0
